==> scripts/02_clean_fdic_sod.R <==
library(tidyverse)
library(janitor)
library(sf)
library(ggmap)
library(mapview)

# download, unzip, rename fdic bank file
download.file("https://www5.fdic.gov/sod/download/ALL_2017_10032017.ZIP",
              "./data/fdic_sod.zip")
unzip("./data/fdic_sod.zip", files = "ALL_2017.csv", exdir = "./data")
file.rename("./data/ALL_2017.csv", "./data/fdic_sod_2017.csv")
file.remove("./data/fdic_sod.zip")

# read in fdic
fdic_sod <- read_csv("./data/fdic_sod_2017.csv", col_types = cols(.default = "c"))

# define atlanta counties
atlanta_msa <- c(
  "Barrow",
  "Bartow",
  "Butts",
  "Carroll",
  "Cherokee",
  "Clayton",
  "Cobb",
  "Coweta",
  "Dawson",
  "DeKalb",
  "Douglas",
  "Fayette",
  "Forsyth",
  "Fulton",
  "Gwinnett",
  "Haralson",
  "Heard",
  "Henry",
  "Jasper",
  "Lamar",
  "Meriwether",
  "Morgan",
  "Newton",
  "Paulding",
  "Pickens",
  "Pike",
  "Rockdale",
  "Spalding",
  "Walton")

# clean up names, filter atl banks, select vars, geocode banks
# this will take a while to geocode
fdic_clean <- fdic_sod %>%
  clean_names() %>%
  filter(cntynamb %in% atlanta_msa & stalpbr == "GA") %>%
  select(uninumbr, namefull, namebr, addresbr, citybr, stalpbr, zipbr, cntynamb,
         sims_latitude, sims_longitude, sims_projection, asset, depdom, depsumbr,
         bkclass, bkmo, brnum, brsertyp, cert, charter, clcode, regagnt, specdesc
         ) %>%
  mutate_at(vars(sims_latitude, sims_longitude, asset:depsumbr), parse_number) %>%
  mutate(addr_long = paste0(addresbr, ", ", citybr, ", ", stalpbr, " ", zipbr)) %>%
  mutate_geocode(addr_long)

# make it into a sf
fdic_clean_sf <- fdic_clean %>%
  st_as_sf(
    coords = c("lon", "lat"),
    agr = "constant",
    crs = 4326,
    stringsAsFactors = TRUE,
    remove = FALSE,
    na.fail = TRUE
    ) %>%
  select(-(sims_latitude:sims_projection), -(addr_long:lat)) %>%
  mutate_at(vars(charter, specdesc), str_to_title) 

# check out a map
mapview(fdic_clean_sf)

# write geojson and csv
st_write(fdic_clean_sf, "./output/atl_fdic.geojson", delete_dsn = TRUE)
